Move NVFP4 ONNX regression into CPU export tests
Keep the NVFP4 ONNX exporter regression with the existing CPU Torch ONNX export coverage instead of a standalone test file.
